+++ packData/packData.mjs
@@ -19,6 +19,7 @@ let packs = [
     'cpr-miscellaneous-items',
     'cpr-3rd-party-class-features',
     'cpr-3rd-party-items',
+    'cpr-3rd-party-feats',
     'cpr-miscellaneous'
 ];
 for (let i of packs) {