+++ scripts/macros/actions/stabilize.js
@@ -1,0 +1,45 @@
+import {actorUtils, effectUtils, genericUtils, itemUtils, rollUtils, workflowUtils} from '../../utils.js';
+async function use({trigger, workflow}) {
+    if (workflow.targets.size != 1) return;
+    let targetActor = workflow.targets.first().actor;
+    let dead = effectUtils.getEffectByStatusID(targetActor, 'dead');
+    if (dead) return;
+    let unconscious = effectUtils.getEffectByStatusID(targetActor, 'unconscious');
+    if (!unconscious) return;
+    let roll = await workflow.actor.rollSkill('med');
+    if (roll.total < 10 || !workflow.targets.size) return;
+    let remarkableRecovery = itemUtils.getItemByIdentifier(targetActor, 'remarkableRecovery');
+    if (!remarkableRecovery) {
+        await genericUtils.update(targetActor, {'system.attributes.death.success': 3});
+        let effectData = {
+            name: genericUtils.translate('CHRISPREMADES.Macros.Stabilize.Stabilized'),
+            img: workflow.item.img,
+            origin: workflow.item.uuid,
+            flags: {
+                dae: {
+                    specialDuration: [
+                        'isHealed'
+                    ],
+                    showIcon: true
+                }
+            }
+        };
+        await effectUtils.createEffect(targetActor, effectData);
+        await rollUtils.rollDice('1d4[' + genericUtils.translate('DND5E.TimeHour') + ']', {chatMessage: true, mode: 'blindroll', flavor: genericUtils.translate('CHRISPREMADES.Macros.Stabilize.WakeUp')});
+    } else {
+        await workflowUtils.syntheticItemRoll(remarkableRecovery, [workflow.targets.first()]);
+    }
+}
+export let stabilize = {
+    name: 'Stabilize',
+    version: '1.0.11',
+    midi: {
+        item: [
+            {
+                pass: 'rollFinished',
+                macro: use,
+                priority: 50
+            }
+        ]
+    }
+};