+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.0.12 Release:
+## Update Notes:
+- Additional Compendiums will now sort checked compendia to the top, then sort by priority (thanks CaiusMartius)
+
 # 1.0.11 Release:
 ## New Content:
 - Remarkable Recovery